jobs/extractor/requirements.txt
firebase-admin>=6.0.0
PyMuPDF>=1.24.0
python-pptx>=1.0.0
